Rename "none" driver to "test"
