+++ Godeps/Godeps.json
@@ -16,7 +16,7 @@
 		"github.com/docker/machine/drivers/generic",
 		"github.com/docker/machine/drivers/google",
 		"github.com/docker/machine/drivers/hyperv",
-		"github.com/docker/machine/drivers/none",
+		"github.com/docker/machine/drivers/test",
 		"github.com/docker/machine/drivers/openstack",
 		"github.com/docker/machine/drivers/rackspace",
 		"github.com/docker/machine/drivers/softlayer",
