+++ cmd/machine-driver-test.go
@@ -1,10 +1,10 @@
 package main
 
 import (
-	"github.com/docker/machine/drivers/none"
+	"github.com/docker/machine/drivers/test"
 	"github.com/docker/machine/libmachine/drivers/plugin"
 )
 
 func main() {
-	plugin.RegisterDriver(none.NewDriver("", ""))
+	plugin.RegisterDriver(test.NewDriver("", ""))
 }
