+++ libmachine/host/host_test.go
@@ -3,7 +3,7 @@ package host
 import (
 	"testing"
 
-	_ "github.com/docker/machine/drivers/none"
+	_ "github.com/docker/machine/drivers/test"
 )
 
 func TestValidateHostnameValid(t *testing.T) {
