+++ libmachine/persist/filestore_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/docker/machine/commands/mcndirs"
-	"github.com/docker/machine/drivers/none"
+	"github.com/docker/machine/drivers/test"
 	"github.com/docker/machine/libmachine/host"
 	"github.com/docker/machine/libmachine/hosttest"
 )
@@ -218,10 +218,10 @@ func TestStoreLoad(t *testing.T) {
 		t.Fatal("Expected driver loaded from store to be of type *host.RawDataDriver and it was not")
 	}
 
-	realDriver := none.NewDriver(h.Name, store.Path)
+	realDriver := test.NewDriver(h.Name, store.Path)
 
 	if err := json.Unmarshal(rawDataDriver.Data, &realDriver); err != nil {
-		t.Fatalf("Error unmarshaling rawDataDriver data into concrete 'none' driver: %s", err)
+		t.Fatalf("Error unmarshaling rawDataDriver data into concrete 'test' driver: %s", err)
 	}
 
 	h.Driver = realDriver
